Party Follow fixed!
I've finally found a smooth model for this.  Basically I'm creating a route on the fly, the movement controller uses it just like any other route.

Still needs work for Zone changes.

I still need to go through assist and things like that to refine it and add more options.

git-svn-id: svn://localhost/tmp/testgc/trunk@389 0497ef49-89d3-4428-85ae-dd61597a36b6